requirements.txt
streamlit==1.31.0
Pillow==10.2.0
pytesseract==0.3.10
openpyxl==3.1.2
